+++ CHANGES.rst
@@ -14,6 +14,8 @@ Migrations
 Changes
 ~~~~~~~
 
+- Switch to datadog statsd client library.
+
 - Consider Wifi based query results accurate enough to satisfy queries.
 
 - Stop maintaining separate Python dependency list in setup.py.